Fix color naming and usage:
"android:colorPrimary" is used for the action-/tool bar,
while "android:colorPrimaryDark" is used for the status bar,
so use the correct attribute- and color naming.

"calculator_error_color" is the color used for the animation to the error state,
and the status bar color turnes to this color,
on normal state, the status bar turnes back to the default status bar color,
so the animation to the normal state should use the default status bar color instead of the accent color.